requirements.txt
pytorch
numpy
opencv-contrib-python
pillow